requirements.txt
# requirements.txt
pandas
scikit-learn
pyyaml
tqdm
joblib
matplotlib
seaborn
jupyter